refactor: Implement useUpdateArticle hook
